Updated README to avoid confusion
ditz prompts users for a "Directory to store issues state in" and stores
this in the same file as their username.  The username should not be
tracked in the repo (it should be unique per user), but the directory to
store issues state in should be shared.  Best to simply inform the user
of what to do for the issues state directory.